finish moving code
